Update package name and publish
