+++ .github/workflows/publish.yml
@@ -51,17 +51,6 @@ jobs:
       - name: Grant execute permission for gradlew
         run: chmod +x gradlew
 
-      - name: Decode GPG Key
-        env:
-          GPG_KEY_BASE64: ${{ secrets.GPG_KEY_BASE64 }}
-        run: |
-          echo "$GPG_KEY_BASE64" | base64 --decode > /tmp/secring.gpg
-
-      - name: Import GPG Key
-        run: |
-          gpg --batch --import /tmp/secring.gpg
-          rm /tmp/secring.gpg
-
       - name: Determine version
         id: version
         run: |
@@ -77,22 +66,16 @@ jobs:
 
       - name: Publish to Maven Central
         env:
-          ORG_GRADLE_PROJECT_mavenCentralUsername: ${{ secrets.MAVEN_CENTRAL_USERNAME }}
-          ORG_GRADLE_PROJECT_mavenCentralPassword: ${{ secrets.MAVEN_CENTRAL_PASSWORD }}
-          ORG_GRADLE_PROJECT_signingInMemoryKey: ${{ secrets.GPG_KEY_ARMOR }}
-          ORG_GRADLE_PROJECT_signingInMemoryKeyId: ${{ secrets.GPG_KEY_ID }}
-          ORG_GRADLE_PROJECT_signingInMemoryKeyPassword: ${{ secrets.GPG_KEY_PASSWORD }}
+          ORG_GRADLE_PROJECT_mavenCentralUsername: ${{ secrets.SONATYPE_USERNAME }}
+          ORG_GRADLE_PROJECT_mavenCentralPassword: ${{ secrets.SONATYPE_PASSWORD }}
+          ORG_GRADLE_PROJECT_signingInMemoryKey: ${{ secrets.SIGNING_KEY }}
+          ORG_GRADLE_PROJECT_signingInMemoryKeyId: ${{ secrets.SIGNING_KEY_ID }}
+          ORG_GRADLE_PROJECT_signingInMemoryKeyPassword: ${{ secrets.SIGNING_PASSWORD }}
         run: |
           ./gradlew publishAllPublicationsToMavenCentralRepository \
             -PVERSION_NAME=${{ steps.version.outputs.version }} \
             --no-daemon \
             --no-configuration-cache
-
-      - name: Clean up
-        if: always()
-        run: |
-          rm -f /tmp/secring.gpg
-          gpg --batch --yes --delete-secret-keys $(gpg --list-secret-keys --keyid-format LONG | grep sec | awk '{print $2}' | cut -d'/' -f2) 2>/dev/null || true
 
   # Create GitHub Release artifacts
   release-artifacts:
