+++ gradle.properties
@@ -15,17 +15,17 @@ android.useAndroidX=true
 android.nonTransitiveRClass=true
 
 # Library info - used in publishing
-GROUP=com.samuolis.posthog
+GROUP=io.github.samuolis
 VERSION_NAME=0.1.0-SNAPSHOT
 POM_NAME=PostHog KMP
 POM_DESCRIPTION=Community-driven Kotlin Multiplatform PostHog analytics SDK for Android, iOS, and Web
-POM_URL=https://github.com/nicepics/posthog-kmp
-POM_SCM_URL=https://github.com/nicepics/posthog-kmp
-POM_SCM_CONNECTION=scm:git:git://github.com/nicepics/posthog-kmp.git
-POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/nicepics/posthog-kmp.git
+POM_URL=https://github.com/samuolis/posthog-kmp
+POM_SCM_URL=https://github.com/samuolis/posthog-kmp
+POM_SCM_CONNECTION=scm:git:git://github.com/samuolis/posthog-kmp.git
+POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/samuolis/posthog-kmp.git
 POM_LICENCE_NAME=MIT License
 POM_LICENCE_URL=https://opensource.org/licenses/MIT
 POM_LICENCE_DIST=repo
 POM_DEVELOPER_ID=samuolis
 POM_DEVELOPER_NAME=Lukas Samuolis
-POM_DEVELOPER_URL=https://github.com/nicepics
+POM_DEVELOPER_URL=https://github.com/samuolis
