+++ posthog-kmp/src/wasmJsMain/kotlin/io/github/samuolis/posthog/PostHogContext.wasmJs.kt
@@ -1,0 +1,14 @@
+package io.github.samuolis.posthog
+
+/**
+ * WasmJS PostHog context.
+ * No platform-specific context is required for WebAssembly.
+ */
+public actual class PostHogContext internal constructor(
+    @Suppress("unused") private val unit: Unit = Unit
+)
+
+/**
+ * Creates a PostHogContext for WasmJS.
+ */
+public actual fun PostHogContext(): PostHogContext = PostHogContext(Unit)
