--- sample/src/androidMain/kotlin/com/samuolis/posthog/sample/Platform.android.kt
@@ -1,3 +1,0 @@
-package com.samuolis.posthog.sample
-
-actual fun getPlatformName(): String = "Android"
