+++ sample/src/androidMain/kotlin/io/github/samuolis/posthog/sample/Platform.android.kt
@@ -1,0 +1,3 @@
+package io.github.samuolis.posthog.sample
+
+actual fun getPlatformName(): String = "Android"
