--- sample/src/iosMain/kotlin/com/samuolis/posthog/sample/MainViewController.kt
@@ -1,8 +1,0 @@
-package com.samuolis.posthog.sample
-
-import androidx.compose.ui.window.ComposeUIViewController
-import com.samuolis.posthog.PostHogContext
-
-fun MainViewController() = ComposeUIViewController {
-    App(postHogContext = PostHogContext())
-}
