--- sample/src/iosMain/kotlin/com/samuolis/posthog/sample/Platform.ios.kt
@@ -1,5 +1,0 @@
-package com.samuolis.posthog.sample
-
-import platform.UIKit.UIDevice
-
-actual fun getPlatformName(): String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
