+++ sample/src/iosMain/kotlin/io/github/samuolis/posthog/sample/MainViewController.kt
@@ -1,0 +1,8 @@
+package io.github.samuolis.posthog.sample
+
+import androidx.compose.ui.window.ComposeUIViewController
+import io.github.samuolis.posthog.PostHogContext
+
+fun MainViewController() = ComposeUIViewController {
+    App(postHogContext = PostHogContext())
+}
