+++ sample/src/iosMain/kotlin/io/github/samuolis/posthog/sample/Platform.ios.kt
@@ -1,0 +1,5 @@
+package io.github.samuolis.posthog.sample
+
+import platform.UIKit.UIDevice
+
+actual fun getPlatformName(): String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
