--- sample/src/wasmJsMain/kotlin/com/samuolis/posthog/sample/Main.kt
@@ -1,12 +1,0 @@
-package com.samuolis.posthog.sample
-
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.window.CanvasBasedWindow
-import com.samuolis.posthog.PostHogContext
-
-@OptIn(ExperimentalComposeUiApi::class)
-fun main() {
-    CanvasBasedWindow(canvasElementId = "ComposeTarget") {
-        App(postHogContext = PostHogContext())
-    }
-}
