--- sample/src/wasmJsMain/kotlin/com/samuolis/posthog/sample/Platform.wasmJs.kt
@@ -1,3 +1,0 @@
-package com.samuolis.posthog.sample
-
-actual fun getPlatformName(): String = "Web (WasmJS)"
