+++ sample/src/wasmJsMain/kotlin/io/github/samuolis/posthog/sample/Main.kt
@@ -1,0 +1,12 @@
+package io.github.samuolis.posthog.sample
+
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.window.CanvasBasedWindow
+import io.github.samuolis.posthog.PostHogContext
+
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    CanvasBasedWindow(canvasElementId = "ComposeTarget") {
+        App(postHogContext = PostHogContext())
+    }
+}
