+++ sample/src/wasmJsMain/kotlin/io/github/samuolis/posthog/sample/Platform.wasmJs.kt
@@ -1,0 +1,3 @@
+package io.github.samuolis.posthog.sample
+
+actual fun getPlatformName(): String = "Web (WasmJS)"
